PolicyOps: move delayedBreakPolicy, add method
